requirements.txt
azure-functions
pandas
tables
azure.storage
xlrd